--- src/main/java/com/yapp/artie/global/error/exception/InvalidTokenException.java
@@ -1,8 +1,0 @@
-package com.yapp.artie.global.error.exception;
-
-public class InvalidTokenException extends BusinessException {
-
-  public InvalidTokenException() {
-    super(ErrorCode.USER_TOKEN_ERROR);
-  }
-}
